refactor: 추후 확장성을 위해 constants로 볼 이미지들 이동
